Added lazy loading to dashboard app
Increase the load speed of the app at the cost of having to load the additional layers later on.

Fixed the data explorer tab to display the table correctly

Will explore the ability to load the extra layers in the background as the user is exploring the main layer next.