Incorporate async
App now uses asynchronous processing to try and improve the user experience. The initial basemap with the levee polygon now loads first and quickly. Then, the choropleth maps, followed by the rest of the layers. The whole process still takes a significant amount of time though, e.g., 3 min, so will need to still optimize via simplification of the polygons.